Update Lennox-iComfort-Thermostat.groovy
Add "def updateSetpoint" section to keep thermostatSetpoint value in sync when in heat or cool mode. Without this the Google Home thermostat does not show the current setpoint if in heat or cool mode.

Also includes an info logging capability and several points in the code where messages will be logged if info logging is on.